refactor the app names outside of the AppTag
